taobaoke/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
constantly==15.1.0
crcmod==1.7
cryptography==2.0.3
cssselect==1.0.1
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.6.4
enum34==1.1.6
futures==3.1.1
grpc==0.3.post19
grpcio==1.6.0
grpcio-tools==1.6.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
lxml==4.0.0
msgpack-python==0.4.8
MySQL-python==1.2.5
olefile==0.44
oss2==2.3.1
parsel==1.2.0
Pillow==4.2.1
pkg-resources==0.0.0
protobuf==3.4.0
pyasn1==0.3.5
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.3.0
pytz==2017.2
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
Scrapy==1.4.0
selenium==3.5.0
service-identity==17.0.0
six==1.10.0
topsdk==1.0.0
Twisted==17.5.0
urllib3==1.22
w3lib==1.18.0
xlrd==1.1.0
zope.interface==4.4.2
